% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraSimManhattan.R
\name{bootIntraSim}
\alias{bootIntraSim}
\title{Function to evaluate intra-similarity with bootstrap procedure}
\usage{
bootIntraSim(x, iter = 250, samp = 0.25, boot = 40000)
}
\description{
Function to evaluate intra-similarity with bootstrap procedure
}
