% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapSimper.R
\name{bootstrapSimper}
\alias{bootstrapSimper}
\title{Function to run an iterative SIMPER analysis}
\usage{
bootstrapSimper(x, cl, iter = 100, prop = 0.1, accr = "")
}
\description{
Function to run an iterative SIMPER analysis
}
